+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Added `find_best_n_motifs` method
 - Added more statistics
 
 ## 0.1.0 (2020-12-17)